:sparkles: new api for current employees (#6)
The new API is written in Kotlin. See https://kotlinlang.org/docs/reference/using-maven.html for more details on the maven configuration introduced to make this work.

The new API has better names for fields and includes employees from all countries (instead of just France for the older API). However it does not include employees that are not of type "EMPLOYEE" in Alibeez, which includes top-level executives.

The old API is not removed.